+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The changes are relative to the previous release, unless the baseline is specifi
 ### Changed since 1.4.0
 
 * Fix build with CMake 3.22
+* Update libyuv.cmd/LocalLibyuv.cmake: 6067afde5 (1922)
 
 ### Removed since 1.4.0