+++ CHANGELOG.md
@@ -11,9 +11,9 @@ The changes are relative to the previous release, unless the baseline is specifi
 ### Changed since 1.4.0
 
 * Fix build with CMake 3.22
+* Update aom.cmd/LocalAom.cmake: v3.13.2
 * Update libxml2.cmd/LocalLibXml2.cmake: v2.15.2
 * Update libyuv.cmd/LocalLibyuv.cmake: 6067afde5 (1922)
-* Fix libaom v3.13.1 build with nasm
 * Support long path names in Windows
 
 ### Removed since 1.4.0