+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The changes are relative to the previous release, unless the baseline is specifi
 
 ## [Unreleased]
 
+## [1.4.1] - 2026-03-20
+
 ### Changed since 1.4.0
 
 * Fix build with CMake 3.22
@@ -1354,7 +1356,8 @@ code.
 - Constants `AVIF_VERSION`, `AVIF_VERSION_MAJOR`, `AVIF_VERSION_MINOR`, `AVIF_VERSION_PATCH`
 - `avifVersion()` function
 
-[Unreleased]: https://github.com/AOMediaCodec/libavif/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/AOMediaCodec/libavif/compare/v1.4.1...HEAD
+[1.4.1]: https://github.com/AOMediaCodec/libavif/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/AOMediaCodec/libavif/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/AOMediaCodec/libavif/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/AOMediaCodec/libavif/compare/v1.2.0...v1.2.1